fix react router to display pages when link is clicked
