--- src/components/Missions.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Missions = () => (
-  <section>
-    Missions Missions Missions Missions Missions Missions Missions Missions
-    Missions Missions Missions Missions Missions Missions Missions Missions
-    Missions Missions
-  </section>
-);
-
-export default Missions;
