--- src/components/MyProfile.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const MyProfile = () => (
-  <div>MyProfile MyProfile MyProfile MyProfile MyProfile MyProfile</div>
-);
-
-export default MyProfile;
