--- src/components/Rockets.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-
-const Rockets = () => (
-  <div>
-    Rockets Rockets Rockets Rockets Rockets Rockets Rockets Rockets Rockets
-    Rockets Rockets Rockets Rockets Rockets Rockets
-  </div>
-);
-
-export default Rockets;
