+++ src/index.js
@@ -2,14 +2,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
-import { Provider } from 'react-redux';
 import './index.css';
 import Navbar from './Navbar';
-import MyProfile from './components/MyProfile';
-import Missions from './components/Missions';
-import Rockets from './components/Rockets';
-import store from './redux/store';
-import App from './App';
+import MyProfile from './components/Views/MyProfile';
+import Missions from './components/Views/Missions';
+import Rockets from './components/Views/Rockets';
 
 const router = createBrowserRouter([
   {
@@ -21,11 +18,11 @@ const router = createBrowserRouter([
         element: <MyProfile />,
       },
       {
-        path: 'missions',
+        path: '/missions',
         element: <Missions />,
       },
       {
-        path: '/',
+        path: 'rockets',
         element: <Rockets />,
       },
     ],
@@ -36,8 +33,5 @@ const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
     <RouterProvider router={router} />
-    <Provider store={store}>
-      <App />
-    </Provider>
   </React.StrictMode>,
 );
